daemon: factor out health-related initialization into separate file
The daemon files are huge and the functions within them quite large. Factor out
functionality related to cilium-health intialization into its own function and
related functions into a new file, health.go in the daemon directory.

Signed-off by: Ian Vernon <ian@cilium.io>
